requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
billiard==4.2.1
botocore==1.40.20
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.5
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-crontab==0.7.1
django-environ==0.12.0
django-filter==25.1
django-ip-geolocation==1.6.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
Faker==37.4.2
frozenlist==1.7.0
gql==4.0.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.5.4
multidict==6.6.4
mysqlclient==2.2.7
packaging==25.0
promise==2.3
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
